refactor: split host folder routes
